# schema/openapi.yaml
openapi: "3.0.0"
info:
  version: 2.0.0
  title: Reservation
servers:
  - url: http://localhost:8080/
components:
  schemas:
    Client:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    CreateClientRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string

    Provider:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    CreateProviderRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string

    Availability:
      type: object
      required:
        - provider_id
        - start_time
        - end_time
      properties:
        provider_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time

    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        status:
          type: string
          enum: [reserved, confirmed]
paths:
  /clients:
    post:
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
  /clients/{clientID}:
    get:
      summary: Get client details
      parameters:
        - name: clientID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Client details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'

  /providers:
    post:
      summary: Create a new provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProviderRequest'
      responses:
        '201':
          description: Provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
  /providers/{providerID}:
    get:
      summary: Get provider details
      parameters:
        - name: providerID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Provider details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
  /providers/availability:
    post:
      summary: Submit provider availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Availability'
      responses:
        '201':
          description: Availability created

  /appointments:
    get:
      summary: Get available appointment slots
      parameters:
        - name: providerID
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: date
          description: In the format YYYY-MM-DD
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: A list of available appointment slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'

    post:
      summary: Reserve an appointment slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                  format: uuid
                provider_id:
                  type: string
                  format: uuid
                availability_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Appointment reserved

  /appointments/{appointmentID}/confirm:
    post:
      summary: Confirm a reservation
      parameters:
        - name: appointmentID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reservation confirmed